requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-dotenv==1.4.2
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.12.4
docopt==0.6.2
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
mccabe==0.6.1
pipreqs==0.4.11
psycopg2==2.9.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
yarg==0.1.9
